% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{acf.lag1.alt}
\alias{acf.lag1.alt}
\title{Alternate lag-1 autocorrelation function}
\usage{
acf.lag1.alt(x)
}
\arguments{
\item{x}{numeric vector for the computation}
}
\description{
This function is used to compute lag-1 autocorrelation for the
\href{https://pubmed.ncbi.nlm.nih.gov/27015380/}{SIP method}
}
\keyword{internal}
