% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{acf.lag1}
\alias{acf.lag1}
\title{Compute lag-one autocorrelation}
\usage{
acf.lag1(x)
}
\arguments{
\item{x}{numeric vector for the computation}
}
\value{
a numeric scalar giving the lag-one autocorrelation
}
\description{
Compute lag-one autocorrelation
}
\keyword{internal}
