% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{AG_zero_cross}
\alias{AG_zero_cross}
\title{Zero crossings for ActiGraph data}
\usage{
AG_zero_cross(data)
}
\arguments{
\item{data}{the data to analyze for zero crossings}
}
\description{
Zero crossings for ActiGraph data
}
\keyword{internal}
