% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{AG_zero_prop}
\alias{AG_zero_prop}
\title{Calculate proportion of zeroes in a data stream}
\usage{
AG_zero_prop(data)
}
\arguments{
\item{data}{the data stream to analyze}
}
\description{
Calculate proportion of zeroes in a data stream
}
\keyword{internal}
