% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AP.R
\name{AP.file.reader}
\alias{AP.file.reader}
\title{Read activPAL event files}
\usage{
AP.file.reader(filename)
}
\arguments{
\item{filename}{path to an activPAL event file}
}
\value{
A data frame of activPAL data
}
\description{
Read activPAL event files
}
\examples{
AP.file.reader(
  system.file("sampledata Events.csv", package = "Sojourn")
)
}
