% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIP_Functions.R
\name{combine.sojourns}
\alias{combine.sojourns}
\title{Combine too-short Sojourns}
\usage{
combine.sojourns(durations, short)
}
\arguments{
\item{durations}{durations of a set of Sojourns}

\item{short}{duration cutoff that defines a too-short Sojourn}
}
\description{
Combine too-short Sojourns
}
\keyword{internal}
