% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sojourns.R
\name{combine_sojourns}
\alias{combine_sojourns}
\alias{combine_soj3x}
\alias{update_durations}
\alias{combine.sojourns}
\title{Combine short Sojourns}
\usage{
combine_soj3x(durations, short, sojourns, verbose)

update_durations(durations, sojourns)

combine.sojourns(durations, short)
}
\arguments{
\item{durations}{durations of a set of Sojourns}

\item{short}{minimum length of one Sojourn}

\item{sojourns}{vector of Sojourn assignments}

\item{verbose}{logical. Print updates to console?}
}
\description{
Iteratively combine sojourns that are too short, until all Sojourns meet the
minimum length requirement
}
\keyword{internal}
