% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIP_Functions.R
\name{enhance.actigraph}
\alias{enhance.actigraph}
\title{Combine ActiGraph and activPAL data}
\usage{
enhance.actigraph(ag, ap)
}
\arguments{
\item{ag}{ActiGraph data}

\item{ap}{activPAL data}
}
\description{
Merge data streams for separate monitors in the
\href{https://www.ncbi.nlm.nih.gov/pubmed/27015380}{SIP method}
}
\keyword{internal}
