% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIP_Functions.R
\name{find.transitions}
\alias{find.transitions}
\title{Find transitions}
\usage{
find.transitions(y)
}
\arguments{
\item{y}{the data stream to examine}
}
\description{
Examine a data stream for qualifying transitions according to the
\href{https://pubmed.ncbi.nlm.nih.gov/27015380/}{SIP method}
}
\keyword{internal}
