% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youth_input_demographic.R
\name{input_demographic}
\alias{input_demographic}
\title{Interactively input demographic information}
\usage{
input_demographic()
}
\value{
A data frame containing an ID, attachment site of the monitor (hip or
  wrist), and the participant's sex, age, and BMI
}
\description{
Interactively input demographic information
}
\examples{
if (interactive()) {
  input_demographic()
}
}
