% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messager.R
\name{messager}
\alias{messager}
\title{Function for tidy assembly of text for printing in the console}
\usage{
messager(n, theGrid = NULL, id = "`Unspecified ID`")
}
\arguments{
\item{n}{message number to assemble and return}

\item{theGrid}{a grid entry from the package's internal data}

\item{id}{character scalar giving the participant ID to print}
}
\value{
text to print to the console
}
\description{
Function for tidy assembly of text for printing in the console
}
\keyword{internal}
