% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_transitions.R
\name{predict_transitions}
\alias{predict_transitions}
\title{Predict transitions using a threshold-based approach}
\usage{
predict_transitions(threshold, change_score, refractory_period = 0)
}
\arguments{
\item{threshold}{the change score transition threshold}

\item{change_score}{vector of change scores}

\item{refractory_period}{integer. Minimum allowable number of indices between
predictions}
}
\value{
A vector indicating transition (1) or non-transition (0)
}
\description{
Predict transitions using a threshold-based approach
}
\examples{
x <- seq(0, 1, 0.01)
predict_transitions(0.75, c(x, rev(x)))
}
