% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIP_Functions.R
\name{prep.nnetinputs}
\alias{prep.nnetinputs}
\title{Shape data for input into the neural networks}
\usage{
prep.nnetinputs(ag, sojourns, lag.fun)
}
\arguments{
\item{ag}{ActiGraph and activPAL combined data}

\item{sojourns}{the Sojourns identified on the data}

\item{lag.fun}{function to use for autocorrelations}
}
\description{
Prepare data for use in the neural networks of the
\href{https://pubmed.ncbi.nlm.nih.gov/27015380/}{SIP method}
}
\keyword{internal}
