% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{replace.na}
\alias{replace.na}
\title{Perform mean imputation}
\usage{
replace.na(x)
}
\arguments{
\item{x}{data on which to perform the imputation}
}
\description{
This function is used to perform mean imputation for the
\href{https://pubmed.ncbi.nlm.nih.gov/27015380/}{SIP method}
}
\keyword{internal}
