% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RuLSIF_predict.R
\name{RuLSIF_predict}
\alias{RuLSIF_predict}
\title{Predict transitions using RuLSIF}
\usage{
RuLSIF_predict(threshold, change_score)
}
\arguments{
\item{threshold}{the change score transition threshold}

\item{change_score}{vector of change scores}
}
\value{
A vector indicating transition (1) or non-transition (0)
}
\description{
Predict transitions using RuLSIF
}
\examples{
x <- seq(0, 1, 0.01)
RuLSIF_predict(0.75, c(x, rev(x)))
}
