% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{SIP_ap}
\alias{SIP_ap}
\title{activPAL sample data for exploring Sojourns Including Posture (SIP) method}
\format{
A data frame with 12257 rows and 5 variables:
\describe{
  \item{Time}{POSIX-formatted variable giving the timestamp for each
  observation}
  \item{DataCount}{Integer value giving the index of the sample from
  which the row of data is drawn}
  \item{Interval}{Duration (in seconds) of the interval from one data point
  to the next}
  \item{ActivityCode}{Integer giving the posture activity classification:
  0 is sedentary, 1 is standing, and 2 is stepping}
  \item{CumulativeStepCount}{Integer giving the cumulative step count from
  the start of the file to the current data point.}
  \item{ActivityScore}{Numeric giving MET-hours}
}
}
\usage{
SIP_ap
}
\description{
activPAL sample data for exploring Sojourns Including Posture (SIP) method
}
\keyword{datasets}
