% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIP_Functions.R
\name{SIP_frame_test}
\alias{SIP_frame_test}
\title{Check the formatting of a data frame for use in the SIP pipeline}
\usage{
SIP_frame_test(frame, expected_var_names, expected_classes, verbose)
}
\arguments{
\item{frame}{The data frame to check}

\item{expected_var_names}{character vector. The expected variable names in
\code{frame}}

\item{expected_classes}{character vector. The expected variable classes in
\code{frame}}

\item{verbose}{logical. Print information to console?}
}
\description{
Check the formatting of a data frame for use in the SIP pipeline
}
\keyword{internal}
