% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sojourn.R
\docType{package}
\name{Sojourn}
\alias{Sojourn}
\alias{Sojourn-package}
\title{Sojourn: Apply Sojourn Methods for Processing ActiGraph Accelerometer Data}
\description{
The Sojourn package provides a convenient way to apply the family of Sojourn
methods published in previous works including Lyden K, Keadle S, Staudenmayer
J, & Freedson P (2014) <doi:10.1249/MSS.0b013e3182a42a2d> and Hibbing PR,
Ellingson LD, Dixon PM, & Welk GJ (2018) <doi:10.1249/MSS.0000000000001486>.
}
