% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youth_name_test.R
\name{youth_name_test}
\alias{youth_name_test}
\title{Check monitor data for required demographic variables}
\usage{
youth_name_test(
  AG,
  demo_names = c("id", "Age", "Sex", "BMI"),
  demo_interactive = FALSE,
  fix_sex_var = TRUE,
  ...
)
}
\arguments{
\item{AG}{a data frame of monitor and demographic data}

\item{demo_names}{character vector of required demographic variable names to
look for}

\item{demo_interactive}{logical. Input demographics interactively if missing
variables are identified during format checking?}

\item{fix_sex_var}{logical. Should "Sex" be changed to "SexM" and dummy
coded?}

\item{...}{further arguments passed to svDialogs functions}
}
\value{
A properly-formatted data frame of monitor data
}
\description{
Check monitor data for required demographic variables
}
\keyword{internal}
